spotify: share API structures and helper so tests compile on Windows
