+++ external/spotify/provider.go
@@ -322,81 +322,6 @@ func (p *SpotifyProvider) Playlists() ([]playlist.PlaylistInfo, error) {
 	return slices.Clone(all), nil
 }
 
-type spotifyArtist struct {
-	Name string `json:"name"`
-}
-
-// spotifyItem is a track or podcast episode object from the Spotify Web API.
-// Playlists can hold both; episodes carry a show instead of artists/album.
-type spotifyItem struct {
-	ID      string          `json:"id"`
-	Name    string          `json:"name"`
-	Type    string          `json:"type"` // "track" or "episode"
-	URI     string          `json:"uri"`  // canonical spotify:track:... / spotify:episode:...
-	Artists []spotifyArtist `json:"artists"`
-	Album   struct {
-		Name        string `json:"name"`
-		ReleaseDate string `json:"release_date"`
-	} `json:"album"`
-	Show struct {
-		Name string `json:"name"`
-	} `json:"show"`
-	ReleaseDate  string `json:"release_date"` // episodes carry this at top level
-	DurationMs   int    `json:"duration_ms"`
-	TrackNumber  int    `json:"track_number"`
-	IsPlayable   *bool  `json:"is_playable"`
-	Restrictions struct {
-		Reason string `json:"reason"`
-	} `json:"restrictions"`
-}
-
-// trackFromItem converts a Spotify playlist/library item into a playlist.Track,
-// handling both music tracks and podcast episodes. It uses the canonical uri
-// the API returns (spotify:track:... or spotify:episode:...) as the path, so
-// the player routes episodes to go-librespot's episode metadata path; building
-// "spotify:track:<id>" for an episode makes go-librespot request track metadata
-// for an episode id, which 404s. Episodes carry no artists/album, so the show
-// name fills those slots for display.
-func trackFromItem(t *spotifyItem) playlist.Track {
-	artists := make([]string, len(t.Artists))
-	for i, a := range t.Artists {
-		artists[i] = a.Name
-	}
-	artist := strings.Join(artists, ", ")
-	album := t.Album.Name
-	if t.Type == "episode" {
-		artist = t.Show.Name
-		album = t.Show.Name
-	}
-
-	releaseDate := t.Album.ReleaseDate
-	if releaseDate == "" {
-		releaseDate = t.ReleaseDate
-	}
-	var year int
-	if len(releaseDate) >= 4 {
-		if y, err := strconv.Atoi(releaseDate[:4]); err == nil {
-			year = y
-		}
-	}
-
-	path := t.URI
-	if path == "" {
-		path = fmt.Sprintf("spotify:track:%s", t.ID) // fallback if uri is absent
-	}
-
-	return playlist.Track{
-		Path:         path,
-		Title:        t.Name,
-		Artist:       artist,
-		Album:        album,
-		Year:         year,
-		Stream:       false, // must be false: true causes togglePlayPause to stop+restart instead of pause/resume
-		DurationSecs: t.DurationMs / 1000,
-		TrackNumber:  t.TrackNumber,
-		Unplayable:   (t.IsPlayable != nil && !*t.IsPlayable) || t.Restrictions.Reason != "",
-	}
-}
 
 // Tracks returns all tracks for the given Spotify playlist ID.
 // Track.Path is set to the canonical spotify: URI for the player to resolve.
